+++ src/euroeval/prompt_templates/summarization.py
@@ -14,6 +14,7 @@ from ..languages import (
     FRENCH,
     GERMAN,
     GREEK,
+    HUNGARIAN,
     ICELANDIC,
     ITALIAN,
     LATVIAN,
@@ -70,6 +71,20 @@ SUMM_TEMPLATES: dict["Language", PromptConfig] = {
         default_prompt_template="Έγγραφο: {text}\nΠερίληψη: {target_text}",
         default_instruction_prompt="Έγγραφο: {text}\n\nΓράψτε μια περίληψη του "
         "παραπάνω εγγράφου.",
+        default_prompt_label_mapping=dict(),
+    ),
+    HUNGARIAN: PromptConfig(
+        default_prompt_prefix=(
+            "Az alábbi szövegek "
+            "tartalmazzák az eredeti cikkeket "
+            "és azok összefoglalóit."
+        ),
+        default_prompt_template="Szöveg: {text}\nÖsszefoglaló: {label}",
+        default_instruction_prompt=(
+            "Szöveg: {text}\n\n"
+            "Adjon egy rövid összefoglalót a fenti szövegről.\n"
+            "Összefoglaló: {label}"
+        ),
         default_prompt_label_mapping=dict(),
     ),
     ENGLISH: PromptConfig(