+++ src/scripts/create_multi_wiki_qa.py
@@ -51,6 +51,7 @@ LANGUAGES = [
     "sl",
     "hr",
     "bs",
+    "hu",
 ]