+++ src/scripts/create_winogrande.py
@@ -40,6 +40,7 @@ LANGUAGES = [
     "fi",
     "fr",
     "hr",
+    "hu",
     "it",
     "lt",
     "lv",