+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added support for Hungarian 🇭🇺! This includes the sentiment classification dataset
+  HuSST, the linguistic acceptability dataset ScaLA-hu, the named entity recognition
+  dataset SzegedNER, the reading comprehension dataset MultiWikiQA-hu, the
+  summarisation dataset HunSum, the knowledge dataset MMLU-hu, and the common-sense
+  reasoning dataset Winogrande-hu.
+
 ## [v16.7.1] - 2025-11-18
 
 ### Fixed